requirements.txt
pandas==1.5.2
apache-airflow-providers-postgres==5.4.0
pytest-env==0.8.1
pytest==7.3.2